cmake: add workaround for libstdc++ dependency in IDF 4.0
ESP-IDF v4.0 is compiled with -nostdlib, so the implicit dependency on
libstdc++ doesn't get added. Although libstdc++ will appear somewhere
on the command line due to other components, this isn't reliable since
libc might not appear after it, leading to linking errors.
The only reliable way to link the application with that IDF version is
to add -lstdc++ -lc after every component which uses libstdc++.