examples: allow examples to be installed from the component manager
Previously the examples used EXTRA_COMPONENT_DIRS build system
variable to add esp-dsp component to the build. The variable was set
to "../../../", therefore the examples could only work out of the box
if they were built from inside the esp-dsp repository.

To support the examples to be downloaded individually from the
component manager, we add a idf_component.yml file to each of the
examples. The file specifies the dependency on esp-dsp component and
also supplies the "override_path" value.

This allows both use cases to be supported:

1. When building the examples from esp-dsp Git repository, the build
   system recognizes "override_path" and uses the esp-dsp component
   from the repository itself. Nothing gets downloaded from the
   Component Registry, same as it was before with EXTRA_COMPONENT_DIRS

2. When the example is downloaded from the Component Registry without
   the rest of the project, "override_path" key is removed
   automatically. Therefore the build system downloads esp-bsp
   component from the Registry, placing it into the managed_components
   subdirectory of the project.

This commit also adds "dependencies.lock" and "managed_components" to
.gitignore, since these files may now be created in the example
directories.